// VIS_Application/app/Model/Masters/CompanyRelated/Lookup.ts
export interface ILookup
{
    Id: number;
    Name: string;
    Type:string;
    Value: string;
    FullName: string;
    ColorCode: string;
    Country: string;
    CreatedOn: string;
    CreatedBy: string;
    UpdatedOn: string;
    UpdatedBy: string;
    IsActive: string;
    CompanyId:number;
    EntityMessage: string;
}